Adds test for new Plot overload
This text is ignore since Travis does have pythonnet installed.

+++ Algorithm/QCAlgorithm.Python.cs
@@ -381,7 +381,7 @@ namespace QuantConnect.Algorithm
             using (Py.GIL())
             {
                 var value = ((dynamic)pyObject).Current.Value;
-                Plot(series, value);
+                Plot(series, (decimal)value);
             }
         }
